:art: Move tests: create utils/url_spec.js
refs #6982
